+++ src/main/java/net/silentchaos512/loot/events/ServerEvents.java
@@ -2,10 +2,10 @@ package net.silentchaos512.loot.events;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.ServerPlayerEntity;
+import net.minecraftforge.event.entity.player.PlayerEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.gameevent.PlayerEvent;
 import net.minecraftforge.fml.network.NetworkDirection;
 import net.silentchaos512.loot.TreasureBags;
 import net.silentchaos512.loot.lib.BagTypeManager;